refactor: refactored cloud functions code
