+++ functions/package.json
@@ -21,6 +21,5 @@
         "eslint": "^5.12.0",
         "eslint-plugin-promise": "^4.0.1",
         "firebase-functions-test": "^0.1.6"
-    },
-    "private": true
+    }
 }
